config: Teach Write function how to find a writable file
This falls back to a file we can write to in the event we try to update
a read-only file. Ex:

 /readonly/readonly.toml:
 [foo]
 bar = 1

 /var/sota.toml
 [foo]
 main = 2

We then do a `Write({foo.bar: 3})`. The code will now see that sota.toml is
both more significant in parsing order *and* can be written to so we'll
produce:

/readonly/readonly.toml:
 [foo]
 bar = 1

 /var/sota.toml
 [foo]
 main = 2
 bar = 3

Signed-off-by: Andy Doan <andy@foundries.io>